requirements.txt
streamlit
streamlit-autorefresh
numpy
opencv-python-headless
Pillow
ultralytics